+++ extensions/git-branches-panel/src/git.ts
@@ -2,6 +2,7 @@ export { cleanRepository, getRepoRoot } from './git/shared';
 
 export * from './git/branchGit';
 export * from './git/remoteGit';
+export * from './git/hosting';
 export * from './git/tagGit';
 export * from './git/stashGit';
 export * from './git/worktreeGit';